Emphasize IOptionsMonitor in Options topic
