--- aspnetcore/fundamentals/configuration/options/sample/Models/MyOptionsWithDelegateConfig.cs
@@ -1,16 +1,0 @@
-﻿namespace UsingOptionsSample.Models
-{
-    #region snippet1
-    public class MyOptionsWithDelegateConfig
-    {
-        public MyOptionsWithDelegateConfig()
-        {
-            // Set default value.
-            Option1 = "value1_from_ctor";
-        }
-        
-        public string Option1 { get; set; }
-        public int Option2 { get; set; } = 5;
-    }
-    #endregion
-}
