--- aspnetcore/fundamentals/configuration/options/sample/Models/MySubOptions.cs
@@ -1,17 +1,0 @@
-﻿namespace UsingOptionsSample.Models
-{
-    #region snippet1
-    public class MySubOptions
-    {
-        public MySubOptions()
-        {
-            // Set default values.
-            SubOption1 = "value1_from_ctor";
-            SubOption2 = 5;
-        }
-        
-        public string SubOption1 { get; set; }
-        public int SubOption2 { get; set; }
-    }
-    #endregion
-}
