--- aspnetcore/fundamentals/configuration/options/sample/Pages/_ViewImports.cshtml
@@ -1,4 +1,0 @@
-@using UsingOptionsSample
-@using UsingOptionsSample.Models
-@namespace UsingOptionsSample.Pages
-@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
