+++ aspnetcore/fundamentals/configuration/options/samples/2.x/OptionsSample/Pages/_ViewImports.cshtml
@@ -1,0 +1,4 @@
+@using SampleApp
+@using SampleApp.Models
+@namespace SampleApp.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
