+++ aspnetcore/fundamentals/configuration/options/samples/2.x/OptionsSample/Pages/_ViewStart.cshtml
