+++ aspnetcore/fundamentals/configuration/options/samples/2.x/OptionsSample/appsettings.json
